% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess-breakerofchains.R
\name{ess_break_chains}
\alias{ess_break_chains}
\title{Wrapper around breakerofchains::get_broken_chain()}
\usage{
ess_break_chains(
  doc_lines,
  doc_cursor_line,
  print_result = TRUE,
  assign_result = TRUE
)
}
\arguments{
\item{doc_lines}{an R vector of strings each represneting a line in the file with the chain}

\item{doc_cursor_line}{the current line the cursor is on}
}
\value{
.chain an object storing the output of the chain unless assign_result=FALSE
}
\description{
This is a small wrapper intended to adapt MilesMcBains breakerofchains library for use with
emacs and ESS.
}
\author{
Justin Silverman
}
